@@ -1,4 +1,6 @@
 // Cynhyrchwyd y ffeil hon yn awtomatig. PEIDIWCH Â MODIWL
 // This file is automatically generated. DO NOT EDIT
 
-export function Greet(): Promise<string>;
+export function InitFrontend(): Promise<void>;
+
+export function Refresh(): Promise<void>;
