@@ -17,14 +17,17 @@ func main() {
 
 	// Create application with options
 	err := wails.Run(&options.App{
-		Title:  "usb-tree",
-		Width:  1024,
-		Height: 768,
+		Title:     "usb-tree",
+		Width:     600,
+		Height:    600,
+		MinWidth:  600,
+		MinHeight: 600,
 		AssetServer: &assetserver.Options{
 			Assets: assets,
 		},
 		BackgroundColour: &options.RGBA{R: 27, G: 38, B: 54, A: 1},
 		OnStartup:        app.startup,
+		OnShutdown:       app.Exit,
 		Bind: []interface{}{
 			app,
 		},
